requirements.txt
streamlit==1.24.0
pandas==1.5.3
numpy==1.24.3
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.93.0
python-dotenv==0.21.1
altair==4.2.2
watchdog==3.0.0